+++ codex-rs/code-mode/src/lib.rs
@@ -1,6 +1,7 @@
 mod cell_actor;
 mod runtime;
 mod service;
+mod session_runtime;
 
 pub use codex_code_mode_protocol::*;
 pub use service::CodeModeService;
